Show information saved to CSV
I fixed the show collection to grab the shows by specific cards so the loop worked and then got the csv to save. Artists are encoded with utf-8 because accents were causing issues. I don't know that we need a csv so if that continues to cause issues we can just not save the data